Group "globalComponents" and "pageComponents" under a single folder called "components"
