+++ frontend/src/components/global/chessboards/BotChessboard.tsx
