+++ frontend/src/components/global/modals/LocalGameOverModal.tsx
