+++ frontend/src/components/global/routes/ProtectedRoute.tsx
@@ -1,0 +1,15 @@
+import { Navigate } from "react-router-dom";
+
+import useIsAuthenticated from "../../hooks/useIsAuthenticated.js";
+
+function ProtectedRoute({ children }) {
+    const isAuthenticated = useIsAuthenticated();
+
+    if (isAuthenticated === null) {
+        return <div>Loading ...</div>;
+    }
+
+    return isAuthenticated ? children : <Navigate to={"/login"} />;
+}
+
+export default ProtectedRoute;
