+++ frontend/src/components/page/playBot/BotCard.tsx
