+++ frontend/src/components/page/playBot/BotCategory.tsx
